bugfix (merge r7279 from workingCode into release)
In neutrino events, the perweight of the scattered lepton was different from the perweights of the final state particles in the file 'FinalEvents.dat'.
Fortunately, this did not influnece the calculation of the total inclusive cross section, since after the correction, negative weights may appear.
We thank Shirley Li for pointing us to this bug. 


git-svn-id: http://gibuu.hepforge.org/svn/releases/release2017@7280 fa0d8c7a-6bf9-0310-8a39-84d780902cd4